Establish a successful baseline for CI jobs
Tweaks involved renaming the odl module to avoid dashes. Also, a copy of
the policy.json file was needed to get existing tests to pass. Something
smarter will have to be devised to avoid holding a copy of a file that
ultimately belongs to Neutron core. Missing doc files were also added.

Lastly, prune the requirements to use only the ones that are stricly
required.

Change-Id: I8addb54d519ce1ab4398d32eee1bebf87c5e37b6