Fix the check/merge jobs for ODL
This patch fixes the pluggable devstack for ODL such that ODL can now
run in the OpenStack gate for the check-tempest-dsvm-networking-odl
job. Notably, there was an issue where we were running "bin/karaf" to
start ODL. This started the console in the screen session. However,
somehow screen was sending a "ctrl-d" to karaf, which was causing it
to exit. We now run "bin/start" which starts karaf in daemon mode. We
also explicitly set things like memory for karaf as well.

Fix a few other small nits while in here.

Change-Id: Iedb3dddac433c8dcf7b46f8119f2e14e3f5db2fd
Closes-Bug: #1426212