Correct test_update_port_mac test
Commit [1] and [2] combined to allow the updating of a neutron port's
MAC address. The second commit fixed a bug in the logic for this which
broke the networking-odl unit tests. This commit fixes it by correctly
verifying you cannot update a port's MAC address when using the
networking-odl ML2 MechanismDriver.

[1] cf99eb988e05c56d3842ac388dfe9ada5aa8af86
[2] 89b9e8597816a5ac2f5a48133777c88abb91914e

Change-Id: I9a37dd1f5ef7c43a9b80504fef399c285b01c565