Get upstream CI job working
This commit unwedges the upstream CI job. It does this by both
upping the ODL_BOOT_WAIT and also by ensuring features are
installed in a manner which allows ODL to work on VMs running
in both the HP and RAX clouds. We will also enable debug logs
by default now as well to make debugging easier.

Change-Id: Ibade3f37c53f273a64413bef57061d6eea45d195
Closes-Bug: #1450603
Signed-off-by: Kyle Mestery <mestery@mestery.com>