Major overhaul of plugin.sh and ODL settings
plugin.sh quickly spun out of proportion. This patch tries to modularize
things a bit more by moving/creating functions around.

Furthermore, If we use ODL, and in particular ODL's L3, ensure that certain
variables are set explicitly during the stacking process. This ensure we
minimize the tuning to be done to local.conf.

Change-Id: Iabde30ffff2924e6f304c385223aeb04c0b76144