odl client: gracefully ignore 404 when deleting
When neutron and odl are out of sync, it's possible for neutron to
send delete request to ODL on the resource which ODL doesn't know.
In this case not found error code, 404, can be ignored gracefully
instead of returning an error to user.
This patch depends on ODL neutron northbound fix so that
it can return 404 instead of internal server error.

Change-Id: I2c7e699c227f1fa3251861a2698410666f767701
Closes-Bug: #1489593
(cherry picked from commit be5129e50562b195778385bbf4023ddbec63ef5e)