Reinstate PUBLIC_BRIDGE as a way of adding interface to PUBLIC_BRIDGE
This is a partial revert of the following gerrit:
    https://review.openstack.org/#/c/248835/

With that, we continue to support the usage of PUBLIC_INTERFACE,
so folks that use a local.conf based on Liberty can still use the
old way of adding interface to br-ex.

While the old way will still work, we will keep it undocumented
and display a message to inform user about the expected usage.

Change-Id: I598a3afaa1d5b84fb4707d21bdf406d94aaf0b7f
Signed-off-by: Flavio Fernandes <ffernand@redhat.com>