devstack: make odl logging friendly for gate job
ODL logs are lost by log rotation which makes it difficult/impossible
to know how ODL behaved. And gate job fails to compress the log file of ODL.
With this patch, log rotation is effectively disabled and the log file of
ODL is now compressed by gate job.

The patch changes configurations related to logging of ODL.
- Honor config variables, DEST, LOGDIR, SCREEN_LOGDIR, when
  determining file name for log
- Disable log rotation by setting max file size to large enough
  so that log won't be lost.
  gate job doesn't know about file name of xxx.log.<N>, so it fails to
  compress them.

Closes-Bug: #1531673
Change-Id: I240bb99570e5afaaf75dce6b0fb055b74cf237e6