security group rule: convert unknown protocol name
Neutron security group API accepts name or integer as protocol.
e.g. "tcp" or "6"
Neutron keeps adding protocol names for more human readability.
It's, however, troublesome for OpenDaylight to catch up those protocol
names. So networking-odl should convert protocol name unknown to
OpenDaylight to integer because this conversion can be easily done
in neutron. It's because neutron already knows it.
Although it's possible to always convert protocol name to integer,
this isn't done because of compatibility.

This depends on OpenDaylight patch which was already merged.
https://git.opendaylight.org/gerrit/#/c/35836/

Change-Id: I50a5678d4adbab7ecc5de7a9616612404644de54