bug work around of Java MOXy bug of 475475
On port creation with dns_name: None, null pointer exception occurs
in opendaylight. This is due to the following bug
https://bugs.eclipse.org/bugs/show_bug.cgi?id=475475
EClipsLink MOXy is used as json parser and it wrongly parses json
and stores wrongly null. In order to work around it, networking-odl
shouldn't send unmapped keys whose value is null.

Closes-Bug: #1546848
Change-Id: I1efce0bfb8f943177a7f3e9555f038f13b3ac2a9