Select current java by setting PATH variable
OpenDaylight CI is failing because PATH variable is pointing to
another java than the one in /usr/bin. Therefore using alternative
command is not having expected effect and ODL is launched using
the wrong Java version (version 8 on this case).

To avoid problems of this kind PATH variable has to be fixed
before launching OpenDaylight instead of making required
java version the default one for the system.

Closes-Bug: #1548734
Change-Id: I83cd0136b80ba613dd60607ccd318c2e5470ac2c