ODL v2: Fix multiple updates race
In case there are multiple updates in the journal there is a chance that
the updates will be sent to ODL controller in the wrong order.

The logic in this patch is to handle the updates in the right order
by checking if there are older updates in the queue.
Once an older update has been found the row will go back to the queue
by setting the state back to 'pending'.

Change-Id: I20fd701355327118886cf7e7e9385133c58213af
Closes-Bug: #1561516