ODL v2: Fix delay in sync pending rows
Fixed the delay in sync pending rows, so when "call_thread_on_end"
is called or thread timer is expired the thread will start the sync logic
with no delay.

Instead of starting the sync thread when it stops the thread will start
only once on the class initialization and will wait for event to be set.

Change-Id: I214c9372726e5e8cf333d8736c23332870377cbf
Closes-Bug: #1569923