Complete port details by journal instead of mech driver
Currently port details are completed by the mechanism driver before
being written to the journal. This is necessary for communicating with
the ODL controller and proper handling of journal dependencies.

Instead, port details should be completed by the journal logic which
can also handle a case when there is no call context (this is needed
by the full sync mechanism, and might have other usages in the future).

Partially-implements: blueprint v2-full-sync
Change-Id: If9ed539e3316f465663242824dc9c011f7ab3375