create dir, doc/specs, for spec
From newton cycle, networking-odl project will adopt spec process
in light weight manner to track its tasks and tech details.
This isn't intended to replace Neutron Stadium Evolution[1], but
to complement it to address networking-odl project specific stuff.
Specs are not for users/operators, but for developers. So it's
better to store those files for specs separated. If it's user facing
feature or public API, it should go to neutron spec process.

The full fledged process similar to core neutron wouldn't be
worthwhile because networking-odl is much smaller. And specs for
networking-odl will be stored in networking-odl repo for now as small
start. This might be revised later.
This patch creates doc/specs directory and moves related specs files
under it.

[1] https://review.openstack.org/#/c/312199/ Neutron Stadium Evolution
Change-Id: Ide74a98251078a22d59032010a4e2c06be6afc77