fix _enrich_port() to return the modified "data"
The function _enrich_port() should enrich the port
"data", which is an input.
It currently clones "data" as "new_data",
modifies security group in "new_data", but does nothing with it.
This commit fixes the function to return "new_data", and
use it in the caller function record().
Additionally, the function currently changes "tenant_id"
in "data", instead of in "new_data". This is fixed as well.
Also, there is no need to copy the "network_id", as it
is copied anyway to "new_data".
Last, the unit tests themselves had a bug where there was
mock for plugin.get_security_group. It was never seen, since
the new data from _enrich_port never reached the DB. The
cPickle class can't pickle (serialize) a MagicMock object,
so when fixing _enrich_port(), this MagicMock-ed security
groups attribute generated an error.

Change-Id: I5f459590ce604e3ed2ed9b34129687853b9389f7
Closes-Bug: #1594691