eliminate portbinding by ODL networking topology
Since portbinding logic based networking topology has design flaw
the logic based on pseudo agent based is being introduced.
- https://review.openstack.org/#/c/308031/
Which is agnostic underlying switch technology and fits well with
the design philosophy of networking-odl.
Once it's merged and ready for use, the code of portbinding based on
networking topology will be eliminated

Change-Id: I585f83bffa4dfa4369a809dd6bc49cc396e4dd99
Co-Authored-By: Rajiv Kumar <rajiv.kumar@nectechnologies.in>