devstack: refactor ODL_L3 logic depending on used feature
External bridge behavior was changed in ODL in the new netvirt.
https://git.opendaylight.org/gerrit/#/c/41925

The external interface provided by provider_mappings will be added
to br-int by default, so there will not be an additional bridge.

In addition, ODL is the one creating the bridge and adding the port
to it, there is no reason for a manual fallback of adding the
interface into the bridge.

provider_mappings does not imply the bridge name - only the link
between the physical network (e.g physnet1) and the port name
to be added to the bridge.

Change-Id: Ic45059d672dbb4058ddef90004ef3cef69936d95