remove add/remove router interface to ODL
ODL neutron northbound processes router interface as NOP and ODL openstack
service provide looks port:owner to understand if the port is attached to the
router or not.
So there is no point to send add/remove router interface operation to ODL,
remove the related code.
NOTE: maybe dependency validator needs to be updated so that it knows
the fact that port is added/removed to/from router by the owner of port.

Change-Id: I7fe48aa35822f8ee0aeda58879cc8b546894d05f