secgroup: pushdown default secgroup rules to ODL
When security group is created, corresponding default security group
rules are also created. But there is no corresponding callbacks. So
when security group is created, the special logic to create security
group rules need to be pushed down to ODL.
The modification to v2 driver is a sort of bug work around until the
patch [1] is needed for v2 driver.

[1] https://review.openstack.org/#/c/281693/

Change-Id: I7305d6c970f352bde0111a9a9fe6612bae1ebd5f
Related-bug: #1546910