devstack: make tempest timeout longer
As tempest test is run serially and we're adding scenario tests,
we're occasionally hitting timeout of 120 mins.
For now as work around, make timeout a bit longer to 180 mins.
In near future(Ocata cycle) after migrating to new ODL
netvirt(vpnservice), this parallel execution should be enabled
and this work around should be removed.

Change-Id: I758de9600e18a817cbd2eeeb71a5e4fbee83090b