mech driver v2: build dict for sg on update
precommit callback for security group precommit doesn't pass
complete information. So build it in callback as bug work around.
precoomit callback should be passed full information.
Once it's fixed, this bug work around should be removed.
The original bug report[1] is filed for ODL, but it turned out
the issue resides in neutron and networking-odl

[1] https://bugs.opendaylight.org/show_bug.cgi?id=7256

Change-Id: I15b6f5321cd3e860de7ba50dd5dde8672259c9ba
Related-bug: #1546910