Add a method to query operations
While implementing QoS, we figured out dependency of operation
on network and port can't be completed with existing api, because
QoS policy data does not have networks and ports id which are
attached to it. This patch will add a method which will fetch
all operations on resources in pending or processing state.

So Idea to solve dependency validation for QoS is to fetch all
rows on network and port operations which are in pending or
processing state using method added in this patch and verify
for each row if network data or port data have specific QoS policy
hooked to it.

Change-Id: I4d1e44b18e41ffd80c6ea85000c6a212d3f459b2
Related-Bug: #1654061