using sys.exit(main()) instead of main()
As we known, Exceptions are raised by the sys.exit() function.
When they are not handled, no stack traceback is printed in the
Python interpreter. In this patch we have known main() got return
values(eg:0, 1), but it can't specifies exit status when terminate
the main thread by unusing sys.exit(). So when using sys.exit(main())
instead of main() may be more readable and reasonable.

Change-Id: Ib2b1d8683409290f78d2366dc25d4387a72065bc