devstack: skip several test cases with v2driver and old netvirt
It turned out that test cases with v2driver and old netvirt that is
voting is a bit unstable.
Since legacy netvirt will be deprecated there is no point to work on
it. But new netvirt should be focused. So skip several test cases for
old netvirt and v2driver combination for now to allow patch to be
merged.

Change-Id: Ife59a2181fcc573a03c98c976ec961491fba0146