Don't call journal's run_sync_thread in unit tests
This method takes care of the general journal processing loop using
events and notifications, there's no need or reason to call it from
various unit tests that want the journal to process the entries, instead
the unit tests will now call sync_pending_entries which just processes
the pending entries.

BGPVPN tests had to be fixed not to use mock values in their fake data.
This causes dumpjson to freak out, but that was swallowed by the
exception handler in sync_pending_entries.
However, that exception handling is referencing e.message which is
deprecated and doesn't exist in Python 3 so it cause a ValueError which
used to be swallowed by run_sync_thread but since we're not calling it
any more it actually propagated to the tests and caused them to fail
(the py35 ones).

Change-Id: I2846345e31fe7c6d964bc946c7da48ab64f1ec61