Correct SUPPORTED_RULES of QoS driver
Type of supported_rules should be dict type instead of list, there
will be an exception[1] when associating qos policy to port. This patch
corrects SUPPORTED_RULES. This param type description in DriverBase()
is incorrect and misleading, I submitted a patch in neutron to correct
it as well.
get_description is remnants of old QoS driver, get rid of it.

[1]Callback neutron.services.qos.qos_plugin.QoSPlugin.
_validate_update_port_callback--9223372036847773271 failed with
"'list' object has no attribute 'get'"

Change-Id: I74a3041cd38fdf8d0e17f35fce223393e198b563