Send MAC updates to ODL for SRIOV PFs
When the MAC on a port changes, as is the case for attaching
VM to SRIOV Physical Function(PF), ODL was not being informed.

This change removes the MAC filter on ODL port updates, so
that ODL will be informed if/when MAC changes for a port.

Change-Id: I58eb690eb212a9b98293a2a0d03b4e10ed5210e3
Closes-Bug: 1686799
(cherry picked from commit 9fd3bb17a3490e6acca2f62973876ed529ca0dce)