don't use run_process for odl
The run_process function in devstack is intended for things that
aren't already daemonizing properly, that have some foreground to work
with, potentially putting output onto the console. It was not intended
for services that already have init systems.

Don't use it for odl. It has a start/stop system already, which is
sufficient.

Change-Id: I0ad88d3ab0eb11c17f9b1140a2e4318990c946a8