Retrieve and process port status updates from ODL
This patch utilizes a websocket to receive port
status updates. Ports are initially set to DOWN
and only after a notification is received on the
ws is the port transitioned to ACTIVE.

This patch also adds logic to ensure port status
update logic is employed only if ODL supports it.

Change-Id: Id315e1085db183d4a8966de833224f220eae0961