Allowing lock to be applied per task basis
To make multiple process or threads to execute periodic
tasks parallely, locks are being applied per tasks basis rather
than on the table.

This patch makes sure that a single periodic thread in a process
exist. All the tasks are managed by a single instance
of PeriodicTask class.

If a task is just executed then other tasks can not acquire lock
in time less then periodic time interval. Therefor this prevent
periodic task to be executed by process on the same host or
different host back to back.

Change-Id: Id9054515721ce28a84b3a42d4cece7e3e096ad0c
Partially-Implements: blueprint v2driver-activities-in-pike