Send port[fixed_ips] update to ODL Controller
Based on the user actions (i.e., a new IPv6 subnet is created in an
existing network, a second IPv6 subnet associated to the router etc)
when there is a change in the port[fixed_ips], the information is not
propogated to the ODL controller since n-odl driver is filtering the
fixed_ips element in the port_update event. This is creating issues
for various use-cases. This patch removes the filter and fixes the issue.

Change-Id: Ie7ec5285f261282d88751781b644ca3fdd820bf0
Closes-bug: #1676863
(cherry picked from commit c73ce4405dfbefbcdb7db9179f6a66c8a7d368b5)