Correction in Resource URL Mapping
It is incorrect to use 'qos/policies' for QOS_POLICIES constant.
Rather 'policies' should be used because it represents the
correct mapping. These variables are also used to fetch resources
during full sync. If 'qos/policies' is used then it will lead to
call an invalid method ('get_qos/policies'). The correct method
is 'get_policies'.

Change-Id: I1bf755aa68eedca0f49eda7ddc1950f4db6040e8