Enable full-sync for the bgpvpn, lbaas, qos, sfc, trunk driver
Currently, mentioned drivers does not support full sync. This
patch enables full sync.

Drivers to be synced are dependant on the deployment model using
only some of the drivers therefor a mechanism is required for
registering resources to be synced dynamically.

Now all the drivers maintain list of resources to be synced and
resources are registered during initialization of the driver.

Now resources to be synced depends on run time therefor
ALL_RESOURCES has been moved from the constant.py to the full_sync
file.

Change-Id: I6c8279fdcf54dd5715dc2e55aed7c7824932a41a
Closes-bug: #1701206