Enable members of lbaas to use custom url builder
Member APIs contains /pool/<pool-id>/member/<member-id> which
increased thse size of the column. And current implementation
is not the clean way of implementing member operation. Because
wrong information related to object type is stored in database.

Customied URL builder for lbaas member resource is used to
the discussed issues.

Change-Id: Ic61aaeb8d8d47c75d3d6834c581fb40968145ba2