Update full stack test
The way test have been done in the past was passing
in any generated uuid for instance and fake tap device
for ovs vif plugging and checks if mac address of plugged
port exists in flow, this may not work with new netvirt.

this patch update the test to list ports on bridge and checks
if tap device exists in results.

Change-Id: Ica0004bfd69e3e29c4d528191c42ff8d3f4485a6