pseudo agent port binding: log owner and device_id
When port binding fails, log device_owner and device_id in addition
to the existing ones. They are helpful to determine what port is for.
e.g. vm, dhcp, router etc when diagnosing failure of port binding.
There is a race condition between dhcp port creation and network
deletion because notification to dhcp agent can be delayed.
If dhcp port is to be bound after network is deleted, the binding
fails. It's safe to ignore such a case.

Change-Id: I26e235fa3e0bdab5e3d568a9a0f60a380642c760