Update subport status for trunk ports
Subports status is not updated when the subport are attached/detached
from the trunk. So, its status remains as DOWN from the neutron
perspective even though it is ACTIVE from the ODL side. The main reason
for this is that, when ODL notifies that the subport status is ACTIVE,
neutron fails to set its status to ACTIVE due to subports not being bound
to any hosts.

This patch ensure that the subport status is modified over time and
aligned with the parent port of the trunk. Therefore, it modifies the
subport status based on:
1. If the subports are attached and the parent port is ACTIVE, the
subports status are set to ACTIVE, otherwise it remains as DOWN.
2. If the subports are detached from the trunk, their status is set to
DOWN.
3. If the parent port status changes (from DOWN to ACTIVE or vice
versa), the subports are updated to copy the parent port status.

Closes-Bug: #1707008
Change-Id: I8a0b034ef2eaaeecdc3ff546620fcc6c96e9bcdc
(cherry picked from commit d08034447b144945feeedc5373646ea005bd294c)