Add SIGHUP handling to journal periodic processing
It is a common practice in *NIX to handle SIGHUP to restart a service
gracefully or reload configurations. This allows the worker to be
reloaded and gives way to future enhancement where we can trigger
journal related tasks externally (ie: full sync, maintenance, etc).

Even though signal handling is setup by neutron-lib, without this patch
SIGHUP is not handled at all by this worker. It just results on a crash
of the process. This patch adds the necessary logic to allow to restart
it without crashing the process.

Change-Id: I4117b3901c7687d86dbf00152fc72b496d7cdbc8