Add pidfile to the JournalPeriodicProcessor worker
This will allow to send signals to the worker. As it is right now
sending a SIGHUP would trigger the maintenance task and thus a
full_sync. Something particularly useful for doing new deployments of
ODL. The pidfile will be created under the directory set in 'state_path'
in the neutron configuration.

Change-Id: Ie2b0eadb8b29a4e0534cb8b982cdb19e0bd1e6a9