Fix: tests were no longer avoiding journal calls
When the journal periodic processor was moved to a worker [1] the method
_start_sync_timer was removed leading to V2DriverAdjustment [2] not
longer patching the periodic execution of the journal. However, this is
not needed anymore, as the workers are not started unless specifically
calling start() on them.

[1]: https://review.openstack.org/#/c/486606/
[2]:
https://review.openstack.org/gitweb?p=openstack/networking-odl.git;a=blob;f=networking_odl/tests/functional/base.py;h=be529f2f8e9359741c8a9ffb488ccedf96b232a4;hb=HEAD#l99

Change-Id: Ibea0a31893d3b99478061502b01b8dc290abe3a1
Closes-Bug: #1711081
(cherry picked from commit 60a0178765e4a97562b79c433ebe325e33b62d81)