Fix gate issue: br-int not getting controller
The issue was that devstack/plugin.sh was adding br-int
*after* calling "ovs-vsctl set-manager" thereby causing the
following sequence of events.
1. devstack sets manager and ovs connects to odl
2. netvirt receives notification of the new ovs node and sees
   that there is no br-int
3. netvirt *decides* to configure br-int and adds the controller
   to the bridge. NOTE: AT THIS POINT THE "TRANSACTION" IS STILL
   BEING BUILT, IT HAS NOT BEEN SENT TO OVS YET
4. devstack creates br-int
5. odl sends the above ovsdb transaction containing br-int and it's
   controller setting but...
6. ovsdb-server rejects the transaction due to the fact that br-int
   already exists
The end result is that br-int does not have a controller.

This patch resolves this by simply reversing the order of br-int
creation and set-manager thereby ensuring that br-int exists before
ovs connects to odl. This is a case that odl/netvirt handles just fine.

I tested this twice on the gate in a separate WIP gerrit and both
times br-int was correctly configured with a controller.

Change-Id: I154234cd9a12d68e8cf9f017ea2c474409c44041
(cherry picked from commit 040234c7c96c17514b339ed72d3844aeb22abfa9)