Fixes db.delete_row deletion by row_id
`delete_row(...)` was deleting by `id` which is not existing in the
OpenDaylightJournal model. To solve it, it now deletes by `seqnum` which
is the correct PRIMARY KEY for it.

This was not a problem before, because the function was never called
with that parametrization and also unit tests weren't present for it.
This now adds unit tests to avoid the problem in the future.

Change-Id: I8f6566e4da5f4927bb838d5d698d796ffee1001e
Related-Bug: #1712547