Fix: retry journal.record on dependency deleted
As it stands, between the moment that a dependency is calculated and the
new entry is recorded in the journal, an operation can ocurr that would
make the dependency irrelevant. If that is the case, the operation fails
with a `oslo_db.exception.DBReferenceError` that is not being caught by
`neutron.db.api.is_retriable` and thus ends with an error thrown to the
log and an unexpected result on the caller.

This was modified to catch that exception and then signal `is_retriable`
to retry the operation again, which would calculate the dependencies
again and try to insert again. This is done by raising a
`oslo_db.exception.RetryRequest` on a `DBReferenceError`.

Change-Id: I31c5fd9d903b3853e13c98c76b1f69c4676cad5e
Closes-Bug: #1713447