Remove WebsocketBadStatusException
Turns out this exception is introduced in version 33 of the
websocket library which is not available on some platforms
(rhel 7.x).

Impact on runtime is that if an exception is thrown while connecting
the websocket, the line "except WebsocketBadStatusException" will
raise a ValueError, crash everything, AND mask the initial exception.
This change makes the exception reporting slightly less specific but
it is still totally useable.

see: https://github.com/websocket-client/websocket-client/blob/master/ChangeLog

Also, add info logging of websocket status while we're at it

Fixes-Bug: 1714160
Change-Id: I876a84e479bcca6df051fda609398b6d5db5b8fa