Delete completed rows immediately when retention=0
Completed rows are causing slow downs in journal processing, adjusted
the code so that when retention is set to 0 the rows are deleted
immediately instead of retained.

The default for retention was set to 0 due to the performance
degredation experienced by retaining the completed rows.

Change-Id: Ibb8e06db0f91a0343badf0bf0c036b10fb133a08
Closes-Bug: #1717018
(cherry picked from commit 7167f6c27b5f052c7992ed067f1fe1ab21c4331a)