Add command line tool to analyze logs
This tool can be used to analyze logs and determine the journal's
operation efficiency.
The tool tracks a journal entry's recording and processing to determine
how much time it took since a journal entry was recorded until it was
processed and sent to ODL.

For this purpose, some minor logging was added to the journal code to
log when an entry is recorded and when it was processed.

Change-Id: Ic77762ed29913a3d1109d156ded0f77a8ac4f403