Adding Zuul v3 migrated legacy jobs
This is the first step according to the migration guide. They are added
here in-repo. The next step will be creating the Zuul v3 native jobs
that will be added non-voting until it is understood that they work as
they should, at which point the jobs created here will be dropped in
favor to those.

Change-Id: If8e1cafd534e44835afede132fd8308d89d2b6c6
Needed-By: I411663c2ea93108453d3a7c7184dc753d08328d2