Have create_pending_row return the entry
This change is needed in order to be able to print the entry's ID to the
log once it's created, in order to have better debug-ability of the logs
(we'll be able to correlate entries better if we know the ID and not
only the object ID which can have multiple entries).

A follow up patch makes use of this ability.

Change-Id: I8eb82a5e5c0124c1eaf5fb3cbf32e12fd169e7ff
Related-Bug: #1713487