Make fullstack and functional native Zuul v3 jobs
As part of the migration to Zuul v3 jobs should be converted to native
versions. This patch uses `devstack`'s native job to provide a base
for the functional and fullstack tests.

When backporting this commit a merge conflict ocurred. This was due to a
patch that imported zuul job settings from project-config [1] that was
done directly on the branch.

[1] a14955f4f6263583619e3cac3cc2c04511f12763

Change-Id: I2fb7e7e1a6c79c8e2edb0c0bff795d61881f2cd5
Depends-on: https://review.openstack.org/#/c/539683/
Depends-on: https://review.openstack.org/#/c/539686/
(cherry picked from commit 7cfbf343dce4522ddef59d29611e4f0cc718e4c4)