Add pre_test_grenade_hook
The br-ex is not brought up by ODL, when devstack creates
initial neutron networks, which end up in failure of tempest
tests, due to floating ip inaccessibility. This is work around
which adds a hook to bring the external bridge up, delete initial
networks and re create netwoks at the end of stack.sh. This hook
can be reverted once the br-ex bringup gets in order.

Change-Id: Id335dc34e00c0109efaf191c97c5306cd177ed4a