Design for full sync and recovery of resources
Full sync and recovery is done through maintenance task. Methods
specific to recovery and full sync of a driver is defined in the
driver itself, but default methods are available corresponding
module for full sync and recovery.

This patch defines default methods through base class
and provides a way, to define driver specific methods for full sync
and recovery.

Change-Id: Ibc70dd758b12d09cc29a9c13d60fc27c8254189d
Partial-Closes: #1713697