Reduce timer for periodic task tests
With the timer set to 0.1 the execution of the tests depending on that
timer take almost double the time and unnecesarly. There have been cases
of the tests hitting the their wait timeouts when the system is under
load in the CI. By lowering it to 0 we achieve the same testing but
without the delays, hopefully this will reduce the amount of random
failures.

Change-Id: I61bd4a4309e274546fea2616002c4e0cc8a9e37d