Fix unit tests for ML2 Mechanism driver
The class that tests the ML2 Mechanism driver was relying on a
configuration parameter that was not honored by the `Ml2PluginV2TestCase`
class. As a result, all TestCases that relied on this mechanism were not
instantiating the `opendaylight_v2` driver.

This patch sets the correct variable to allow the parent class to
instantiate the driver. In addition, it adds a test that will allow to
make sure that if there are any changes on the parent class and the
mechanism driver is not instantiated, then it will fail the unit tests.

Change-Id: I4bf2533ba6382a61e6dc010786da1d685a509781