Enabling ML2 for new full sync and recovery mechanism
This patch provide basic code for enabling full sync and recovery
for ML2. It keeps existing mechanism as it is, but implements
other protocols expected by new full sync mechanism.

This patch adds resources to driver classes, which will be used
by base driver to register for full sync and recovery.

Change-Id: I00b5a5f927ef3b2cce1bfa203a1402dd3cd1c61e
Partial-Bug: #1713697