Pass binding:profile attribute as a string
Opendaylight defines the "binding:profile" attribute as a string
(see [1]), so in order to parse it we need to send it as one.
Openstack defines binding:profile as a dict that can hold several data
type values e.g string or list of strings therefore is sent as a string
and parsed in odl for specific attributes by their type.

[1] https://github.com/opendaylight/neutron/blob/7dd96afe5bee07c5e1b8f4aa2bfd5d25193de684/model/src/main/yang/neutron-binding.yang#L64

Change-Id: I560a70eb94495d466b03271d67e9c9191bac0a0e