Remove unit tests for ML2 Mechanism driver
The test cases importing `neutron.tests.unit.plugins.ml2.test_plugin` and
creating classes with noop operation [1] follow a faulted pattern. The
reasons are twofold:

1. If the mechanism driver is not loaded correctly [2], the tests then
don't test the mechanism driver but still succeed and therefore there is
no indication that there is something wrong with the code.

2. It gives a false sense of reassurance. If the code of those tests is
analyzed it's possible to see that the code itself is mostly centered
around testing the REST endpoint of neutron than actually testing that
the mechanism succeeds on the operation it was supposed to test. As a
result of this, there is marginally added value on having those tests.
To be clear, the hooks for the respective operations are called on the
mechanism driver, but the result of the operation is not asserted.

For more reference please refer to the mailing list conversations that
discuss this issue:

1. https://lists.opendaylight.org/pipermail/neutron-dev/2017-December/001499.html
2. http://lists.openstack.org/pipermail/openstack-dev/2017-December/125466.html

[1]: http://codesearch.openstack.org/?q=class%20.*\(.*TestMl2
[2]: something that was happening in networking-odl and addressed by
https://review.openstack.org/#/c/523934

Change-Id: I242a7f69d62f6e3365164cf8d104a5cf0a8499b0