Fix dependency calculation when two fixed IPs under same subnet
Two fixed IPs could be assigned for one port under a same subnet. In
this case there is a bug that calculates dependencies wrongly and
therefore causes the creation of the journal entry failed because there
are duplicate dependencies.

This patch makes sure that when there are duplicate subnets then only
one of those dependencies is returned.

Change-Id: I70e4759e153ae58c0428832940d21ba67c215b7b
Closes-Bug: #1742402