Log exceptions on security group callbacks
This patch adds EXCEPTION-level logging for non-retriable exceptions
during security groups and rules callbacks. The reason is that the
callback manager triggering these callbacks wont do any logging on its
own for precommit events. On situations like [1], it was necessary to
enable debug logging to check the reason behind the error.

Regarding [1], see change [2] for a fix that prevents the exception.

[1] http://jira.opendaylight.org/browse/NETVIRT-896
[2] https://review.openstack.org/#/c/536935

Change-Id: I49872b5b65372ac333ad759bb506fac7348c3fa9
co-authored-by: Jaime Caamaño Ruiz <jcaamano@suse.com>