Fix missing variables in devstack/override-defaults
The mechanism to override defaults in devstack is called very early in
the stack process, even before calling the plugin.sh of the plugin that
is being loaded. This causes an issue, since the default settings of ODL
live in devstack/settings.odl but they are not imported by the time the
override mechanism is called and therefore it doesn't execute correctly.

This patch makes V2 of the driver as the default choice unless otherwise
specified.

Change-Id: I1da6d6111b713c38fc4a7460fc72981faab1a10d