Improve IPC and forking reliability in tests
These tests were randomly timing out in the CI:

  - JournalPeriodicProcessorTest.test_handle_sighup_gracefully
  - JournalPeriodicProcessorTest.test_reset_called_on_sighup

These failures are suspected to have been happening due to a race
condition between the test and the child process. This patch adds more
robustness to the way the pipe is used, refactors their code for better
maintainability and adds a forced delay to give time for the child
process to settle before starting to test.

Change-Id: I1a90963d1d30e8c03b2842fda8e1a77dd2c687fe